+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Add optional `compact_str` feature; `smartstring` remains supported for now and will be removed in a future release. The features are mutually exclusive (#29).
+
 ## 0.5.0
 * Bump `rand` to 0.9
 * Add `rngs::thread_local` random source (#36)